refactor: navigator wrapper for tests
