+++ src/screens/Bookmarks/Navbar.spec.tsx
@@ -1,23 +1,9 @@
-import { NavigationContainer } from '@react-navigation/native'
-import { createStackNavigator } from '@react-navigation/stack'
 import { render } from '@testing-library/react-native'
-import React, { ReactNode } from 'react'
+import React from 'react'
 
-import withContexts from '../../components/with-contexts'
+import wrapper from '../../lib/NavigatorContext'
 
 import Navbar from './Navbar'
-
-const { Screen, Navigator } = createStackNavigator()
-
-type WrapperProps = { children: ReactNode }
-
-const wrapper = ( { children }: WrapperProps ) => withContexts(
-  <NavigationContainer>
-    <Navigator>
-      <Screen name="default">{() => children}</Screen>
-    </Navigator>
-  </NavigationContainer>,
-)
 
 describe( '<Navbar />', () => {
   it( 'should render a back button', () => {
